Framework for GameDirector added; created new class to handle all responsibility related to each players' character
