+++ NightFight/Assets/CharacterDataManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d7fea933145b84c59bb2d520305d596a
-timeCreated: 1477266933
+guid: 8da0ba0ed33aa4c9bb12ac121972dce3
+timeCreated: 1477266880
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
